+++ src/eventsub/channel/mod.rs
@@ -71,6 +71,8 @@ pub use subscribe::{ChannelSubscribeV1, ChannelSubscribeV1Payload};
 #[doc(inline)]
 pub use subscription::{ChannelSubscriptionEndV1, ChannelSubscriptionEndV1Payload};
 #[doc(inline)]
+pub use subscription::{ChannelSubscriptionGiftV1, ChannelSubscriptionGiftV1Payload};
+#[doc(inline)]
 pub use unban::{ChannelUnbanV1, ChannelUnbanV1Payload};
 #[doc(inline)]
 pub use update::{ChannelUpdateV1, ChannelUpdateV1Payload};